#15: Move all biometrics-related code to biometrics.dart.
